pyproject.toml
[tool.poetry]
name = "routine-tracker"
version = "0.1.0"
description = ""
authors = ["Denys Mitlenko <mitlenkodenis3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-htmx = "^1.19.0"
psycopg2-binary = "^2.9.9"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
django-components = "^0.102"
django-widget-tweaks = "^1.5.0"
pytz = "^2024.2"
django-webpush = "^0.3.6"
django-celery-beat = "^2.7.0"
celery = "^5.4.0"
redis = "^5.2.0"
gunicorn = "^23.0.0"
django-markdownfield = "^0.11.0"


[tool.poetry.group.dev.dependencies]
colorlog = "^6.8.2"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
black = "^24.10.0"
faker = "^30.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.pyi?$'
force-exclude = '''
/(
  | migrations
)/
'''
